requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
alembic==1.8.1
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.2
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.3
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
black==23.9.1
bleach==6.0.0
blinker==1.6.2
bokeh==3.2.0
cassandra-driver==3.28.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-params==0.3.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
dask==2023.7.0
dataclasses-json==0.6.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distributed==2023.7.0
distro==1.8.0
docker==6.1.3
ecdsa==0.18.0
executing==1.2.0
fastapi==0.99.1
fastapi-utils==0.2.1
fastjsonschema==2.17.1
flake8==6.1.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.0.5
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
geomet==0.2.1.post1
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.2
httplib2==0.19.1
httptools==0.6.0
httpx==0.25.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==7.8.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.5
jupyterlab_server==2.23.0
kiwisolver==1.4.4
llama-index==0.9.10
locket==1.0.0
lz4==4.3.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.8
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.1
openai==1.3.7
orjson==3.8.14
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
passlib==1.7.4
pathspec==0.11.2
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.1
plotly==5.15.0
prometheus-client==0.17.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==1.10.12
pyflakes==3.1.0
Pygments==2.15.1
PyMySQL==1.0.3
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-socketio==5.8.0
python-terraform==0.10.1
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.0
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.38.3
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
stack-data==0.6.2
starlette==0.27.0
tblib==2.0.0
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tiktoken==0.5.1
tinycss2==1.2.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.17.6
validators==0.18.2
watchgod==0.8.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
widgetsnbextension==3.6.5
wrapt==1.16.0
xyzservices==2023.5.0
yarl==1.9.3
zenml==0.44.0
zict==3.0.0
zipp==3.16.1
